requirements.txt
alabaster==0.7.12
anaconda-client==1.2.2
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
beautifulsoup4==4.8.2
bootstrap4==0.1.0
certifi==2019.11.28
chardet==3.0.4
clyent==1.2.1
colorama==0.4.3
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-datetimepicker==3.14
django-crispy-forms==1.8.1
django-datetimepicker==3.14
django-heroku==0.3.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.4
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.0.0
